Moved StyledTooltip to SnapshotMap/styles.js
