+++ src/components/SnapshotMap/index.js
@@ -8,7 +8,6 @@ import {
 } from "react-simple-maps";
 import SnapshotMapLegend from "../SnapshotMapLegend";
 import MapZoomButtons from "../MapZoomButtons";
-import { StyledTooltip } from "../styled-components/MapTooltip";
 import SnapshotMapCountryPopup from "../SnapshotMapCountryPopup";
 import {
   getEbolaFillColorsDictionary,
@@ -17,7 +16,7 @@ import {
 } from "../../utils/snapshotMapHelpers";
 import { getCountryDiseaseCountDictionary } from "../../utils/snapshotMapHelpers";
 import { countriesCoordinates } from "../../constants/CountriesCoordinates";
-import { SnapshotMapContainer } from "./styles";
+import { SnapshotMapContainer, StyledTooltip } from "./styles";
 
 const SnapshotMap = ({
   filters,
